fix: 상태 변경 핸들러에서 파라미터에 따라 분기 하도록 수정

+++ src/main/java/egovframework/com/cop/smt/sdm/web/EgovDeptSchdulManageController.java
@@ -16,7 +16,6 @@ import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.multipart.MultipartHttpServletRequest;
@@ -37,6 +36,7 @@ import egovframework.com.cop.smt.sdm.service.DeptSchdulManageVO;
 import egovframework.com.cop.smt.sdm.service.EgovDeptSchdulManageService;
 import egovframework.com.utl.fcc.service.EgovStringUtil;
 import jakarta.annotation.Resource;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 
 /**
@@ -499,16 +499,17 @@ public class EgovDeptSchdulManageController {
 	 * @return "egovframework/com/cop/smt/sdm/EgovDeptSchdulManageDetail"
 	 * @throws Exception
 	 */
-	@RequestMapping(value = "/cop/smt/sdm/EgovDeptSchdulManageDetail.do", method = RequestMethod.POST)
+	@RequestMapping(value = "/cop/smt/sdm/EgovDeptSchdulManageDetail.do")
 	public String egovDeptSchdulManageDetail(@ModelAttribute("searchVO") ComDefaultVO searchVO,
-			DeptSchdulManageVO deptSchdulManageVO, @RequestParam Map<String, String> commandMap, ModelMap model)
+			DeptSchdulManageVO deptSchdulManageVO, @RequestParam Map<String, String> commandMap, ModelMap model,
+			HttpServletRequest request)
 			throws Exception {
 
 		String sLocationUrl = "egovframework/com/cop/smt/sdm/EgovDeptSchdulManageDetail";
 
 		String sCmd = commandMap.get("cmd");
 
-		if ("del".equals(sCmd)) {
+		if ("del".equals(sCmd) && "POST".equalsIgnoreCase(request.getMethod())) {
 			egovDeptSchdulManageService.deleteDeptSchdulManage(deptSchdulManageVO);
 			sLocationUrl = "redirect:/cop/smt/sdm/EgovDeptSchdulManageList.do";
 		} else {
